+++ src/ts/Ads/Views/Privacy/IPrivacySDKViewHandler.ts
@@ -1,0 +1,10 @@
+import { IPrivacySettings } from 'Privacy/IPrivacySettings';
+
+export interface IPrivacySDKViewHandler {
+    onPrivacyReady(): void;
+    onPrivacyCompleted(userSettings: IPrivacySettings): void;
+    onPrivacyOpenUrl(url: string): void;
+    onPrivacyMetric(data: { [key: string]: unknown }): void;
+    onPrivacyFetch(url: string, data: { [key: string]: unknown }): void;
+    onPrivacyViewError(event: string | Event): void;
+}