+++ src/ts/Core/ICore.ts
@@ -36,6 +36,7 @@ import { StorageBridge } from 'Core/Utilities/StorageBridge';
 import { NativeErrorApi } from 'Core/Api/NativeErrorApi';
 import { DeviceIdManager } from 'Core/Managers/DeviceIdManager';
 import { ClassDetectionApi } from 'Core/Native/ClassDetection';
+import { SdkDetectionInfo } from 'Core/Models/SdkDetectionInfo';
 
 export interface ICoreApi extends IModuleApi {
     Cache: CacheApi;
@@ -82,4 +83,5 @@ export interface ICore extends IApiModule {
     DeviceInfo: DeviceInfo;
     Config: CoreConfiguration;
     Ads: IAds;
+    SdkDetectionInfo: SdkDetectionInfo;
 }